Added new base class and new Object Toggle and LightToggle
New Base class for handling DayTimeUpdater and Enviro

New EnviroLights to toggle lights on / off if the time of day is day or night.

New EnviroObject toggle to toggle visibility of object if the time of day is day or night.
This class has a two properties for if the object is enabled during the day or night or both.